fix: correct navigation
