+++ app/components/create-account/initialFormData.js
@@ -1,0 +1,13 @@
+export const initialFormData = {
+    firstName: '',
+    lastName: '',
+    email: '',
+    profession: '',
+    tel: '',
+    businessName: '',
+    businessEmail: '',
+    businessTel: '',
+    staffs: '',
+    password: '',
+    retypePassword: ''
+}
